+++ library/src/main/java/com/pengrad/telegrambot/request/AddStickerToSet.java
@@ -9,15 +9,15 @@ import com.pengrad.telegrambot.response.BaseResponse;
  */
 public class AddStickerToSet extends AbstractUploadRequest<AddStickerToSet, BaseResponse> {
 
-    public static AddStickerToSet tgsSticker(Integer userId, String name, String emojis, Object tgsSticker) {
+    public static AddStickerToSet tgsSticker(Long userId, String name, String emojis, Object tgsSticker) {
         return new AddStickerToSet(userId, name, emojis, "tgs_sticker", tgsSticker);
     }
 
-    public AddStickerToSet(Integer userId, String name, Object pngSticker, String emojis) {
+    public AddStickerToSet(Long userId, String name, Object pngSticker, String emojis) {
         this(userId, name, emojis, "png_sticker", pngSticker);
     }
 
-    private AddStickerToSet(Integer userId, String name, String emojis, String stickerParam, Object sticker) {
+    private AddStickerToSet(Long userId, String name, String emojis, String stickerParam, Object sticker) {
         super(BaseResponse.class, stickerParam, sticker);
         add("user_id", userId);
         add("name", name);
