+++ library/src/main/java/com/pengrad/telegrambot/request/GetChatMember.java
@@ -8,7 +8,7 @@ import com.pengrad.telegrambot.response.GetChatMemberResponse;
  */
 public class GetChatMember extends BaseRequest<GetChatMember, GetChatMemberResponse> {
 
-    public GetChatMember(Object chatId, int userId) {
+    public GetChatMember(Object chatId, long userId) {
         super(GetChatMemberResponse.class);
         add("chat_id", chatId).add("user_id", userId);
     }
