+++ library/src/main/java/com/pengrad/telegrambot/request/GetGameHighScores.java
@@ -8,12 +8,12 @@ import com.pengrad.telegrambot.response.GetGameHighScoresResponse;
  */
 public class GetGameHighScores extends BaseRequest<GetGameHighScores, GetGameHighScoresResponse> {
 
-    public GetGameHighScores(int userId, Object chatId, int messageId) {
+    public GetGameHighScores(long userId, Object chatId, int messageId) {
         super(GetGameHighScoresResponse.class);
         add("user_id", userId).add("chat_id", chatId).add("message_id", messageId);
     }
 
-    public GetGameHighScores(int userId, String inlineMessageId) {
+    public GetGameHighScores(long userId, String inlineMessageId) {
         super(GetGameHighScoresResponse.class);
         add("user_id", userId).add("inline_message_id", inlineMessageId);
     }
