+++ library/src/main/java/com/pengrad/telegrambot/request/RestrictChatMember.java
@@ -9,12 +9,12 @@ import com.pengrad.telegrambot.response.BaseResponse;
  */
 public class RestrictChatMember extends BaseRequest<RestrictChatMember, BaseResponse> {
 
-    public RestrictChatMember(Object chatId, int userId) {
+    public RestrictChatMember(Object chatId, long userId) {
         super(BaseResponse.class);
         add("chat_id", chatId).add("user_id", userId);
     }
 
-    public RestrictChatMember(Object chatId, int userId, ChatPermissions permissions) {
+    public RestrictChatMember(Object chatId, long userId, ChatPermissions permissions) {
         super(BaseResponse.class);
         add("chat_id", chatId).add("user_id", userId).add("permissions", permissions);
     }
