+++ library/src/main/java/com/pengrad/telegrambot/request/SetGameScore.java
@@ -9,12 +9,12 @@ import com.pengrad.telegrambot.response.SendResponse;
  */
 public class SetGameScore extends BaseRequest<SetGameScore, BaseResponse> {
 
-    public SetGameScore(int userId, int score, Object chatId, int messageId) {
+    public SetGameScore(long userId, int score, Object chatId, int messageId) {
         super(SendResponse.class);
         add("user_id", userId).add("score", score).add("chat_id", chatId).add("message_id", messageId);
     }
 
-    public SetGameScore(int userId, int score, String inlineMessageId) {
+    public SetGameScore(long userId, int score, String inlineMessageId) {
         super(BaseResponse.class);
         add("user_id", userId).add("score", score).add("inline_message_id", inlineMessageId);
     }
