+++ library/src/main/java/com/pengrad/telegrambot/request/UploadStickerFile.java
@@ -8,7 +8,7 @@ import com.pengrad.telegrambot.response.GetFileResponse;
  */
 public class UploadStickerFile extends AbstractUploadRequest<UploadStickerFile, GetFileResponse> {
 
-    public UploadStickerFile(Integer userId, Object pngSticker) {
+    public UploadStickerFile(Long userId, Object pngSticker) {
         super(GetFileResponse.class, "png_sticker", pngSticker);
         add("user_id", userId);
     }
